cleaned up configuration code. added newe module - SettingUtils.py
